Major refactor for simplicity
