+++ relod/envs/base_reacher_env.py
@@ -3,20 +3,20 @@
 #
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
-import numpy as np
 import time
 import gym
-import sys
-from multiprocessing import Array, Value
 
+import numpy as np
+import cv2 as cv
+
+from multiprocessing import Array, Value
 from senseact.rtrl_base_env import RTRLBaseEnv
 from senseact.devices.ur import ur_utils
-from envs.visual_ur5_reacher.ur_setup import setups
+from relod.envs.ur5_setup import setups
 from senseact.sharedbuffer import SharedBuffer
 from senseact import utils
-import cv2 as cv
-from envs.visual_ur5_reacher.camera_communicator import CameraCommunicator, DEFAULT_HEIGHT, DEFAULT_WIDTH
-from envs.visual_ur5_reacher.monitor_communicator import MonitorCommunicator
+from relod.envs.camera_communicator import CameraCommunicator, DEFAULT_HEIGHT, DEFAULT_WIDTH
+
 
 class ReacherEnv(RTRLBaseEnv, gym.core.Env):
     """A class implementing Visual-UR5 Reaching and tracking environments.
@@ -504,13 +504,13 @@ class ReacherEnv(RTRLBaseEnv, gym.core.Env):
                 self._pstop_time_ = time.time()
                 self._pstop_times_.append(self._pstop_time_)
                 # Check to see if too many p-stops occurred within a short time window
-                if len(self._pstop_times_) > self._max_pstop:
-                    if self._pstop_time_ - self._pstop_times_[-self._max_pstop] < self._max_pstop_window:
-                        print("Too many p-stops encountered, closing environment")
-                        print('Greater than {0} p-stops encountered within {1} seconds'.format(
-                                    self._max_pstop, self._max_pstop_window))
-                        self.close()
-                        sys.exit(1)
+                # if len(self._pstop_times_) > self._max_pstop:
+                #     if self._pstop_time_ - self._pstop_times_[-self._max_pstop] < self._max_pstop_window:
+                #         print("Too many p-stops encountered, closing environment")
+                #         print('Greater than {0} p-stops encountered within {1} seconds'.format(
+                #                     self._max_pstop, self._max_pstop_window))
+                #         self.close()
+                #         sys.exit(1)
             elif time.time() > self._pstop_time_ + self._clear_pstop_after:  # XXX
                 print("Unlocking p-stop")
                 self._actuation_packet_['UR5'] = self._pstop_unlock_packet
@@ -752,36 +752,11 @@ class ReacherEnv(RTRLBaseEnv, gym.core.Env):
         return inside_bound, inside_buffer_bound, mat, xyz
 
     def _compute_reward_(self, image, joint):
-        """Computes reward at a given time step.
-        Returns:
-            A float reward.
-        """
-        image = image[:, :, -3:]
-        lower = [0, 0, 120]
-        upper = [50, 50, 255]
-        lower = np.array(lower, dtype="uint8")
-        upper = np.array(upper, dtype="uint8")
+        return -1
 
-        mask = cv.inRange(image, lower, upper)
-        size_x, size_y = mask.shape
-        # reward for reaching task, may not be suitable for tracking
-        if 255 in mask:
-            xs, ys = np.where(mask == 255.)
-            reward_x = 1 / 2  - np.abs(xs - int(size_x / 2)) / size_x
-            reward_y = 1 / 2 - np.abs(ys - int(size_y / 2)) / size_y
-            reward = np.sum(reward_x * reward_y) / self._image_width / self._image_height
-        else:
-            reward = 0
-        reward *= 800
-        reward = np.clip(reward, 0, 4)
+    def imm_stop_arm(self):
 
-        '''
-        When the joint 4 is perpendicular to the mounting ground:
-            joint 0 + joint 4 == 0
-            joint 1 + joint 2 + joint 3 == -pi
-        '''
-        scale = (np.abs(joint[0] + joint[4]) + np.abs(np.pi + np.sum(joint[1:4])))
-        return reward - scale
+        self._actuator_comms['UR5'].actuator_buffer.write(self._stopj_packet)
 
     def _check_done(self):
         """Checks whether the episode is over.
@@ -795,10 +770,13 @@ class ReacherEnv(RTRLBaseEnv, gym.core.Env):
         """
         self._episode_steps += 1
         if (self._episode_steps >= self._episode_length_step): # or env_done:
-            self._actuator_comms['UR5'].actuator_buffer.write(self._stopj_packet)
+            self.stop_arm()
             return True
         else:
             return False
+
+    def stop_arm(self):
+        self._actuator_comms['UR5'].actuator_buffer.write(self._stopj_packet)
 
     def reset(self, blocking=True):
         """Resets the arm, optionally blocks the environment until done."""
