--- relod/envs/ur5_config.py
@@ -1,20 +1,0 @@
-
-import numpy as np
-
-config = {
-    'conv': [
-        # in_channel, out_channel, kernel_size, stride
-        [-1, 32, 3, 2],
-        [32, 32, 3, 2],
-        [32, 32, 3, 2],
-        [32, 32, 3, 1],
-    ],
-
-    #'latent': 50,
-
-    'mlp': [
-        [-1, 1024],
-        [1024, 1024],
-        [1024, -1]
-    ],
-}
